requirements.txt
aiohttp
names
argon2-cffi
